Superstore Sales Dashboard Video
This is actually a common practice dataset of an e-commerce business. But I have pretended it is from Walmart and used the Walmart logo and theme to show the use of brand colours in the theme. The video shows a quick demonstration of the dashboard with the slicer moving to different years. All the cards and graphs change accordingly.